+++ packages/react/src/themes/base/components/Provider/providerBoxVariables.ts
@@ -1,0 +1,9 @@
+export interface ProviderBoxVariables {
+  background: string
+  color: string
+}
+
+export default (siteVariables): ProviderBoxVariables => ({
+  background: siteVariables.bodyBackground,
+  color: siteVariables.bodyColor,
+})